requirements.txt
Flask==1.1.1
Flask-MySQLdb==1.0.1
gunicorn
WTForms==2.3.3
passlib
wtforms-html5==0.6.1
email_validator
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
tensorflow==2.4.1
pillow==7.2.0
opencv-python>=4.1.2
absl-py~=0.10
appdirs==1.4.3
astor==0.8.1
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
distlib==0.3.0
filelock==3.0.12
gast==0.3.3
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio~=1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
numpy
oauthlib==3.1.0
opt-einsum==3.3.0
pickleshare==0.7.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.5.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
virtualenv==20.0.17
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0